requirements.txt
streamlit
google-generativeai
langchain
langchain-community
langchain-google-genai
PyPDF2
faiss-cpu
python-dotenv
